# docs/technical-specification/architecture/openapi.yaml
openapi: 3.1.0
info:
  title: Hive Chameleon Client API
  version: 0.1.0
  summary: Client-facing HTTP contract for the Unity clients and Next.js portal.
  description: |
    Draft client API for Hive Chameleon. Realtime lobby, matchmaking, and gameplay
    messages are owned by Nakama and are documented in the paired architecture
    document rather than this OpenAPI file.

    Operations tagged `x-maturity: conceptual` describe a future API shape only.
    They do not approve unresolved product, custody, payment-provider, resale,
    organizer, or gameplay rules.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Relative client API base; deployment hostname is configuration.
...
security:
  - bearerAuth: []
paths:
  /auth/hive/challenges:
    post:
      tags: [Authentication]
      operationId: createHiveLoginChallenge
      summary: Create a single-use direct-Hive login challenge
      x-priority: P0
      x-maturity: committed
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HiveChallengeRequest'
      responses:
        '201':
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HiveChallenge'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        '503':
          $ref: '#/components/responses/ServiceUnavailableProblem'
  /auth/hive/sessions:
    post:
      tags: [Authentication]
      operationId: createHiveSession
      summary: Verify a Hive challenge signature and establish a game session
      x-priority: P0
      x-maturity: committed
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HiveSessionRequest'
      responses:
        '201':
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '503':
          $ref: '#/components/responses/ServiceUnavailableProblem'
  /auth/guest:
    post:
      tags: [Authentication]
      operationId: createGuestSession
      summary: Establish an isolated guest session for the public WebGL client
      description: |
        Available only in explicitly enabled public-game environments. Guest identities
        cannot sign Hive operations, use custody, or impersonate durable Hive players.
      x-priority: P0
      x-maturity: committed
      security: []
      responses:
        '201':
          description: Isolated guest session
          headers:
            Cache-Control:
              description: Prevent storage of browser session credentials by HTTP caches
              required: true
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '503':
          $ref: '#/components/responses/ServiceUnavailableProblem'
  /auth/google/exchange:
    post:
      tags: [Authentication, Onboarding]
      operationId: exchangeGoogleAuthorization
      summary: Exchange a validated Google OIDC authorization result
      description: |
        Google authenticates the game session; it is not a Hive signature. A known
        identity receives normal game tokens. A first-time identity receives a
        restricted onboarding token and next-step state without creating a player.
      x-priority: P0
      x-maturity: committed
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleExchangeRequest'
      responses:
        '200':
          description: Existing player session or pending onboarding session
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AuthTokens'
                  - $ref: '#/components/schemas/OnboardingSession'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '503':
          $ref: '#/components/responses/ServiceUnavailableProblem'
  /auth/refresh:
    post:
      tags: [Sessions]
      operationId: refreshSession
      summary: Rotate the refresh credential and issue a new access token
      x-priority: P0
      x-maturity: committed
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Rotated credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
  /auth/session:
    delete:
      tags: [Sessions]
      operationId: revokeCurrentSession
      summary: Revoke the current game session
      x-priority: P0
      x-maturity: committed
      responses:
        '204':
          description: Session revoked
        '401':
          $ref: '#/components/responses/UnauthorizedProblem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived game API access token.
    onboardingBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Restricted token scoped to one verified Google external identity and onboarding job.
...
  responses:
...
    UnauthorizedProblem:
      description: Authentication is missing, invalid, expired, or revoked
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ConflictProblem:
      description: Domain state conflict or duplicate logical command
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationProblem:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailableProblem:
      description: A required Hive, signer, custody, HAF, or game dependency is unavailable
      headers:
        Retry-After:
          description: Suggested retry delay in seconds.
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      example:
        type: /problems/domain-conflict
        title: Domain state conflict
        status: 409
        code: lobby_state_conflict
        detail: The lobby changed before the command was applied.
        requestId: 01900000-0000-7000-8000-000000000001
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[a-z0-9_]+$'
        detail:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    HiveUsername:
      type: string
      minLength: 3
      maxLength: 16
      pattern: '^[a-z][a-z0-9.-]*[a-z0-9]$'
    HiveChallengeRequest:
      type: object
      additionalProperties: false
      required: [hiveUsername, platform]
      properties:
        hiveUsername:
          $ref: '#/components/schemas/HiveUsername'
        platform:
          $ref: '#/components/schemas/ClientPlatform'
    HiveChallenge:
      type: object
      additionalProperties: false
      required: [id, hiveUsername, challenge, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        hiveUsername:
          $ref: '#/components/schemas/HiveUsername'
        challenge:
          type: string
          description: Canonical backend-generated challenge text, not a secret.
        expiresAt:
          type: string
          format: date-time
    HiveSessionRequest:
      type: object
      additionalProperties: false
      required: [challengeId, hiveUsername, signingProvider, signature]
      properties:
        challengeId:
          type: string
          format: uuid
        hiveUsername:
          $ref: '#/components/schemas/HiveUsername'
        signingProvider:
          $ref: '#/components/schemas/ExternalSigningProvider'
        signature:
          type: string
          maxLength: 512
    GoogleExchangeRequest:
      type: object
      additionalProperties: false
      required: [authorizationCode, codeVerifier, redirectUri, platform]
      properties:
        authorizationCode:
          type: string
          maxLength: 4096
          writeOnly: true
        codeVerifier:
          type: string
          minLength: 43
          maxLength: 128
          writeOnly: true
        redirectUri:
          type: string
          format: uri
          maxLength: 2048
        platform:
          $ref: '#/components/schemas/ClientPlatform'
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 32
          maxLength: 2048
          writeOnly: true
    AuthTokens:
      type: object
      additionalProperties: false
      required: [kind, accessToken, refreshToken, accessTokenExpiresAt, sessionId, player]
      properties:
        kind:
          type: string
          const: authenticated
        accessToken:
          type: string
          format: password
          x-sensitive: true
        refreshToken:
          type: string
          format: password
          x-sensitive: true
        accessTokenExpiresAt:
          type: string
          format: date-time
        sessionId:
          type: string
          format: uuid
        player:
          $ref: '#/components/schemas/PlayerSummary'
    OnboardingSession:
      type: object
      additionalProperties: false
      required: [kind, onboardingToken, externalIdentityId, nextStep, expiresAt]
      properties:
        kind:
          type: string
          const: onboarding
        onboardingToken:
          type: string
          format: password
          x-sensitive: true
        externalIdentityId:
          type: string
          format: uuid
        nextStep:
          type: string
          enum: [choose_username, provisioning_pending, provisioning_ready]
        provisioningId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
...
    PlayerSummary:
      type: object
      additionalProperties: false
      required: [id, hiveUsername]
      properties:
        id:
          type: string
          format: uuid
        hiveUsername:
          $ref: '#/components/schemas/HiveUsername'
        displayName:
          type: string
          maxLength: 128
        avatarUrl:
          type: string
          format: uri
...
    ClientPlatform:
      type: string
      enum: [webgl, windows, macos, linux]
    ExternalSigningProvider:
      type: string
      enum: [keychain, hiveauth, hivesigner, other_external]
...